Extracted the parsing of the configuration file from the Retriever module and put it in a new module named Config. This is done because an other new module named Db also needs the configuration file.
De file Db.py is new and contains a function that interfaces to a MySQL 
database. This database is used for the mapping of NM to NP accession numbers.

The file mutalyzer.conf now has two more options: a MySQL user name and a 
database name. 

Started the documentation of a fresh installation, see Install.txt for more
details.

In Mutator.py: Added a function that calculates the positions of splice sites
after mutatations.

In Scheduler.py: Added some code (still in comment) that can parallelise jobs
by using treads and detecting the number of processors present on the host.



git-svn-id: https://humgenprojects.lumc.nl/svn/mutalyzer/trunk@5 eb6bd6ab-9ccd-42b9-aceb-e2899b4a52f1