Deletions of entire exons (even number of consecutive splice sites).
This merges the first part of exon-deletions-branch (the second part concerns
unknown exon offsets, denoted with the ? character). Except that we now keep
track of removed splice sites in the transcript, instead of in the Mutator
instance. I guess the latter was better, so we will probably revert to storing
this in the Mutator instance.

Issue #35.



git-svn-id: https://humgenprojects.lumc.nl/svn/mutalyzer/branches/refactor-mutalyzer-branch@233 eb6bd6ab-9ccd-42b9-aceb-e2899b4a52f1