Use implicit configuration object creation
The old way of using the configuration file was by instantiating a Config
object which read the file. This instance was passed to every function and
object that might need it.

The new way is by simply calling config.get('name') to get the configuration
value for 'name'. This lazily reads the configuration file and the contents
are cached for future calls.


git-svn-id: https://humgenprojects.lumc.nl/svn/mutalyzer/branches/implicit-config-branch@408 eb6bd6ab-9ccd-42b9-aceb-e2899b4a52f1